Move UtilToLLVMPass to common place.
