+++ mlir/include/mlir-extensions/Conversion/util_to_llvm.hpp
@@ -1,0 +1,35 @@
+// Copyright 2022 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <functional>
+#include <memory>
+
+namespace mlir {
+class LowerToLLVMOptions;
+class MLIRContext;
+class Pass;
+} // namespace mlir
+
+namespace plier {
+
+/// Convert operations from the plier_util dialect to the LLVM dialect.
+///
+/// TODO: We cannot pass LowerToLLVMOptions directly to the pass as it requires
+/// mlir context which is not yet available at this point, pass creation
+/// function instead.
+std::unique_ptr<mlir::Pass> createUtilToLLVMPass(
+    std::function<mlir::LowerToLLVMOptions(mlir::MLIRContext &)> optsGetter);
+} // namespace plier
