+++ numba_dpcomp/numba_dpcomp/mlir_compiler/lib/pipelines/lower_to_llvm.cpp
@@ -52,6 +52,7 @@
 #include "base_pipeline.hpp"
 #include "pipelines/plier_to_std.hpp"
 
+#include "mlir-extensions/Conversion/util_to_llvm.hpp"
 #include "mlir-extensions/Dialect/plier/dialect.hpp"
 #include "mlir-extensions/Dialect/plier_util/dialect.hpp"
 #include "mlir-extensions/Transforms/func_utils.hpp"
@@ -816,81 +817,6 @@ struct DeallocOpLowering
   }
 };
 
-struct ExpandShapeLowering
-    : public mlir::ConvertOpToLLVMPattern<mlir::memref::ExpandShapeOp> {
-  using ConvertOpToLLVMPattern<
-      mlir::memref::ExpandShapeOp>::ConvertOpToLLVMPattern;
-
-  explicit ExpandShapeLowering(mlir::LLVMTypeConverter &converter)
-      : ConvertOpToLLVMPattern<mlir::memref::ExpandShapeOp>(converter) {}
-
-  mlir::LogicalResult
-  matchAndRewrite(mlir::memref::ExpandShapeOp op,
-                  mlir::memref::ExpandShapeOp::Adaptor adaptor,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto dstType = op.getType().cast<mlir::MemRefType>();
-    auto converter = getTypeConverter();
-    auto dstLlvmType = converter->convertType(dstType);
-    if (!dstLlvmType)
-      return mlir::failure();
-
-    auto inds = op.getReassociationIndices();
-
-    int64_t currentDstInd = 0;
-    auto dstShape = dstType.getShape();
-    for (auto &ind : inds) {
-      int64_t reassocInd = -1;
-      for (auto i : ind) {
-        assert(i >= 0);
-        if (dstShape[currentDstInd] != 1) {
-          if (reassocInd != -1)
-            return mlir::failure();
-          reassocInd = i;
-        }
-        ++currentDstInd;
-      }
-    }
-
-    mlir::MemRefDescriptor src(adaptor.src());
-
-    auto loc = op->getLoc();
-    auto result = mlir::MemRefDescriptor::undef(rewriter, loc, dstLlvmType);
-    result.setAllocatedPtr(rewriter, loc, src.allocatedPtr(rewriter, loc));
-    result.setAlignedPtr(rewriter, loc, src.alignedPtr(rewriter, loc));
-    result.setOffset(rewriter, loc, src.offset(rewriter, loc));
-
-    auto indexType = getIndexType();
-    auto zero = rewriter.create<mlir::LLVM::ConstantOp>(
-        loc, indexType, rewriter.getIntegerAttr(indexType, 0));
-    auto one = rewriter.create<mlir::LLVM::ConstantOp>(
-        loc, indexType, rewriter.getIntegerAttr(indexType, 1));
-
-    currentDstInd = 0;
-    for (auto it : llvm::enumerate(inds)) {
-      auto &ind = it.value();
-      auto srcInd = it.index();
-      for (auto i : ind) {
-        assert(i >= 0);
-        mlir::Value dim;
-        mlir::Value stride;
-        if (dstShape[currentDstInd] != 1) {
-          dim = src.size(rewriter, loc, srcInd);
-          stride = src.stride(rewriter, loc, srcInd);
-        } else {
-          dim = one;
-          stride = zero;
-        }
-        result.setSize(rewriter, loc, currentDstInd, dim);
-        result.setStride(rewriter, loc, currentDstInd, stride);
-        ++currentDstInd;
-      }
-    }
-
-    rewriter.replaceOp(op, static_cast<mlir::Value>(result));
-    return mlir::success();
-  }
-};
-
 class LLVMFunctionPass : public mlir::OperationPass<mlir::LLVM::LLVMFuncOp> {
 public:
   using OperationPass<mlir::LLVM::LLVMFuncOp>::OperationPass;
@@ -1275,477 +1201,6 @@ struct PostLLVMLowering
   }
 };
 
-struct LowerRetainOp : public mlir::ConvertOpToLLVMPattern<plier::RetainOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  mlir::LogicalResult
-  matchAndRewrite(plier::RetainOp op, plier::RetainOp::Adaptor adaptor,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto arg = adaptor.source();
-    if (!arg.getType().isa<mlir::LLVM::LLVMStructType>())
-      return mlir::failure();
-
-    auto llvmVoidPointerType = getVoidPtrType();
-    auto incref_func = [&]() {
-      auto mod = op->getParentOfType<mlir::ModuleOp>();
-      assert(mod);
-      auto func = mod.lookupSymbol<mlir::LLVM::LLVMFuncOp>("NRT_incref");
-      if (!func) {
-        mlir::OpBuilder::InsertionGuard guard(rewriter);
-        rewriter.setInsertionPointToStart(mod.getBody());
-        auto llvmVoidType = getVoidType();
-        func = rewriter.create<mlir::LLVM::LLVMFuncOp>(
-            rewriter.getUnknownLoc(), "NRT_incref",
-            mlir::LLVM::LLVMFunctionType::get(llvmVoidType,
-                                              llvmVoidPointerType));
-      }
-      return func;
-    }();
-
-    mlir::MemRefDescriptor source(arg);
-
-    auto loc = op.getLoc();
-    mlir::Value ptr = source.allocatedPtr(rewriter, loc);
-    ptr = rewriter.create<mlir::LLVM::BitcastOp>(loc, llvmVoidPointerType, ptr);
-    rewriter.create<mlir::LLVM::CallOp>(loc, incref_func, ptr);
-    rewriter.replaceOp(op, arg);
-
-    return mlir::success();
-  }
-};
-
-struct LowerExtractMemrefMetadataOp
-    : public mlir::ConvertOpToLLVMPattern<plier::ExtractMemrefMetadataOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  mlir::LogicalResult
-  matchAndRewrite(plier::ExtractMemrefMetadataOp op,
-                  plier::ExtractMemrefMetadataOp::Adaptor adaptor,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto arg = adaptor.source();
-    if (!arg.getType().isa<mlir::LLVM::LLVMStructType>())
-      return mlir::failure();
-
-    auto loc = op.getLoc();
-    mlir::MemRefDescriptor src(arg);
-    auto val = [&]() -> mlir::Value {
-      auto index = op.dimIndex().getSExtValue();
-      if (index == -1)
-        return src.offset(rewriter, loc);
-
-      return src.stride(rewriter, loc, static_cast<unsigned>(index));
-    }();
-
-    rewriter.replaceOp(op, static_cast<mlir::Value>(val));
-    return mlir::success();
-  }
-};
-
-struct LowerBuildTuple
-    : public mlir::ConvertOpToLLVMPattern<plier::BuildTupleOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  mlir::LogicalResult
-  matchAndRewrite(plier::BuildTupleOp op, plier::BuildTupleOp::Adaptor adaptor,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto converter = getTypeConverter();
-    auto type = converter->convertType(op.getType());
-    if (!type)
-      return mlir::failure();
-
-    auto loc = op.getLoc();
-    mlir::Value init = rewriter.create<mlir::LLVM::UndefOp>(loc, type);
-    for (auto it : llvm::enumerate(adaptor.args())) {
-      auto arg = it.value();
-      auto newType = arg.getType();
-      assert(newType);
-      auto casted = doCast(rewriter, loc, arg, newType);
-      auto index = rewriter.getI64ArrayAttr(static_cast<int64_t>(it.index()));
-      init =
-          rewriter.create<mlir::LLVM::InsertValueOp>(loc, init, casted, index);
-    }
-
-    rewriter.replaceOp(op, init);
-    return mlir::success();
-  }
-};
-
-struct LowerUndef : public mlir::ConvertOpToLLVMPattern<plier::UndefOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  mlir::LogicalResult
-  matchAndRewrite(plier::UndefOp op, plier::UndefOp::Adaptor /*adaptor*/,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto converter = getTypeConverter();
-    auto type = converter->convertType(op.getType());
-    if (!type)
-      return mlir::failure();
-
-    rewriter.replaceOpWithNewOp<mlir::LLVM::UndefOp>(op, type);
-    return mlir::success();
-  }
-};
-
-static void addToGlobalDtors(mlir::ConversionPatternRewriter &rewriter,
-                             mlir::ModuleOp mod, mlir::SymbolRefAttr attr,
-                             int32_t priority) {
-  auto loc = mod->getLoc();
-  auto dtorOps = mod.getOps<mlir::LLVM::GlobalDtorsOp>();
-  auto prioAttr = rewriter.getI32IntegerAttr(priority);
-  mlir::OpBuilder::InsertionGuard g(rewriter);
-  if (dtorOps.empty()) {
-    rewriter.setInsertionPoint(mod.getBody(), std::prev(mod.getBody()->end()));
-    auto syms = rewriter.getArrayAttr(attr);
-    auto priorities = rewriter.getArrayAttr(prioAttr);
-    rewriter.create<mlir::LLVM::GlobalDtorsOp>(loc, syms, priorities);
-    return;
-  }
-  assert(llvm::hasSingleElement(dtorOps));
-  auto dtorOp = *dtorOps.begin();
-
-  auto addpendArray = [&](mlir::ArrayAttr arr,
-                          mlir::Attribute attr) -> mlir::ArrayAttr {
-    auto vals = arr.getValue();
-    llvm::SmallVector<mlir::Attribute> ret(vals.begin(), vals.end());
-    ret.emplace_back(attr);
-    return rewriter.getArrayAttr(ret);
-  };
-  auto newDtors = addpendArray(dtorOp.getDtors(), attr);
-  auto newPrioritiess = addpendArray(dtorOp.getPriorities(), prioAttr);
-  rewriter.setInsertionPoint(dtorOp);
-  rewriter.create<mlir::LLVM::GlobalDtorsOp>(loc, newDtors, newPrioritiess);
-  rewriter.eraseOp(dtorOp);
-}
-
-struct LowerTakeContextOp
-    : public mlir::ConvertOpToLLVMPattern<plier::TakeContextOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  mlir::LogicalResult
-  matchAndRewrite(plier::TakeContextOp op,
-                  plier::TakeContextOp::Adaptor adaptor,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto converter = getTypeConverter();
-    assert(converter);
-    auto ctx = op.context();
-    auto ctxType = converter->convertType(ctx.getType());
-    if (!ctxType)
-      return mlir::failure();
-
-    mlir::ValueRange results = op.results();
-    auto resultsCount = static_cast<unsigned>(results.size());
-    llvm::SmallVector<mlir::Type> resultTypes(resultsCount);
-    for (auto i : llvm::seq(0u, resultsCount)) {
-      auto type = converter->convertType(results[i].getType());
-      if (!type)
-        return mlir::failure();
-
-      resultTypes[i] = type;
-    }
-
-    auto ctxStructType =
-        mlir::LLVM::LLVMStructType::getLiteral(getContext(), resultTypes);
-    auto ctxStructPtrType = mlir::LLVM::LLVMPointerType::get(ctxStructType);
-
-    auto mod = op->getParentOfType<mlir::ModuleOp>();
-    assert(mod);
-
-    auto unknownLoc = rewriter.getUnknownLoc();
-    auto loc = op->getLoc();
-    auto wrapperType =
-        mlir::LLVM::LLVMFunctionType::get(getVoidType(), ctxType);
-    auto wrapperPtrType = mlir::LLVM::LLVMPointerType::get(wrapperType);
-    mlir::Value initFuncPtr;
-
-    auto insertFunc = [&](mlir::StringRef name, mlir::Type type,
-                          mlir::LLVM::Linkage linkage) {
-      mlir::OpBuilder::InsertionGuard g(rewriter);
-      rewriter.setInsertionPointToStart(mod.getBody());
-      return rewriter.create<mlir::LLVM::LLVMFuncOp>(unknownLoc, name, type,
-                                                     linkage);
-    };
-
-    auto lookupFunc = [&](mlir::StringRef name, mlir::Type type) {
-      // TODO: fix and use lookupOrCreateFn
-      if (auto func = mod.lookupSymbol<mlir::LLVM::LLVMFuncOp>(name))
-        return func;
-
-      mlir::OpBuilder::InsertionGuard g(rewriter);
-      rewriter.setInsertionPointToStart(mod.getBody());
-      return rewriter.create<mlir::LLVM::LLVMFuncOp>(
-          unknownLoc, name, type, mlir::LLVM::Linkage::External);
-    };
-
-    if (auto initFuncSym = adaptor.initFuncAttr()) {
-      auto funcName = initFuncSym.getLeafReference().getValue();
-      auto wrapperName = (funcName + "_wrapper").str();
-
-      auto initFunc = [&]() {
-        mlir::OpBuilder::InsertionGuard g(rewriter);
-        auto func =
-            insertFunc(wrapperName, wrapperType, mlir::LLVM::Linkage::Private);
-        auto block = rewriter.createBlock(
-            &func.getBody(), mlir::Region::iterator{}, ctxType, unknownLoc);
-        rewriter.setInsertionPointToStart(block);
-
-        // Get init func declaration so we can check original return types.
-        auto initFunc = mod.lookupSymbol<mlir::func::FuncOp>(initFuncSym);
-        assert(initFunc && "Invalid init func");
-        auto initFuncType = initFunc.getFunctionType();
-        assert(initFuncType.getNumResults() == resultsCount &&
-               "Invalid init func");
-
-        auto innerResults =
-            rewriter
-                .create<mlir::func::CallOp>(unknownLoc, initFuncSym,
-                                            initFuncType.getResults())
-                ->getResults();
-
-        mlir::Value ctxStruct =
-            rewriter.create<mlir::LLVM::UndefOp>(unknownLoc, ctxStructType);
-        for (auto i : llvm::seq(0u, resultsCount)) {
-          auto pos = rewriter.getI64ArrayAttr(i);
-          auto srcType = initFuncType.getResult(i);
-          auto convertedType = converter->convertType(srcType);
-          assert(convertedType && "Invalid init func result type");
-
-          mlir::Value val = innerResults[i];
-          // Init function may not be type-converted at this point, so insert
-          // conversion casts.
-          if (convertedType != srcType)
-            val = converter->materializeSourceConversion(rewriter, unknownLoc,
-                                                         convertedType, val);
-          assert(val && "Invalid init func result type");
-
-          ctxStruct = rewriter.create<mlir::LLVM::InsertValueOp>(
-              unknownLoc, ctxStruct, val, pos);
-        }
-        auto ptr = rewriter.create<mlir::LLVM::BitcastOp>(
-            unknownLoc, ctxStructPtrType, block->getArgument(0));
-        rewriter.create<mlir::LLVM::StoreOp>(unknownLoc, ctxStruct, ptr);
-        rewriter.create<mlir::LLVM::ReturnOp>(unknownLoc, llvm::None);
-        return func;
-      }();
-
-      initFuncPtr = rewriter.create<mlir::LLVM::AddressOfOp>(loc, initFunc);
-    } else {
-      initFuncPtr = rewriter.create<mlir::LLVM::NullOp>(loc, wrapperPtrType);
-    }
-
-    mlir::Value deinitFuncPtr;
-    if (auto deinitFuncSym = adaptor.releaseFuncAttr()) {
-      auto funcName = deinitFuncSym.getLeafReference().getValue();
-      auto wrapperName = (funcName + "_wrapper").str();
-
-      auto deinitFunc = [&]() {
-        mlir::OpBuilder::InsertionGuard g(rewriter);
-        auto func =
-            insertFunc(wrapperName, wrapperType, mlir::LLVM::Linkage::Private);
-        auto block = rewriter.createBlock(
-            &func.getBody(), mlir::Region::iterator{}, ctxType, unknownLoc);
-        rewriter.setInsertionPointToStart(block);
-
-        auto ptr = rewriter.create<mlir::LLVM::BitcastOp>(
-            unknownLoc, ctxStructPtrType, block->getArgument(0));
-        auto ctxStruct =
-            rewriter.create<mlir::LLVM::LoadOp>(unknownLoc, ctxStructType, ptr);
-
-        // Get deinit func declaration so we can check original arg types.
-        auto deinitFunc = mod.lookupSymbol<mlir::func::FuncOp>(deinitFuncSym);
-        assert(deinitFunc && "Invalid deinit func");
-        auto deinitFuncType = deinitFunc.getFunctionType();
-        assert(deinitFuncType.getNumInputs() == resultsCount);
-
-        llvm::SmallVector<mlir::Value> args(resultsCount);
-        for (auto i : llvm::seq(0u, resultsCount)) {
-          auto pos = rewriter.getI64ArrayAttr(i);
-          mlir::Value val = rewriter.create<mlir::LLVM::ExtractValueOp>(
-              unknownLoc, resultTypes[i], ctxStruct, pos);
-          auto resType = deinitFuncType.getInput(i);
-          // Deinit function may not be type-converted at this point, so insert
-          // conversion casts.
-          if (resultTypes[i] != resType)
-            val = converter->materializeTargetConversion(rewriter, unknownLoc,
-                                                         resType, val);
-
-          args[i] = val;
-        }
-
-        rewriter.create<mlir::func::CallOp>(unknownLoc, deinitFuncSym,
-                                            llvm::None, args);
-        rewriter.create<mlir::LLVM::ReturnOp>(unknownLoc, llvm::None);
-        return func;
-      }();
-
-      deinitFuncPtr = rewriter.create<mlir::LLVM::AddressOfOp>(loc, deinitFunc);
-    } else {
-      deinitFuncPtr = rewriter.create<mlir::LLVM::NullOp>(loc, wrapperPtrType);
-    }
-
-    auto takeCtxFunc = [&]() -> mlir::LLVM::LLVMFuncOp {
-      llvm::StringRef name("dpcompTakeContext");
-      auto retType = getVoidPtrType();
-      const mlir::Type argTypes[] = {
-          mlir::LLVM::LLVMPointerType::get(getVoidPtrType()),
-          getIndexType(),
-          wrapperPtrType,
-          wrapperPtrType,
-      };
-      auto funcType = mlir::LLVM::LLVMFunctionType::get(retType, argTypes);
-      return lookupFunc(name, funcType);
-    }();
-
-    auto purgeCtxFunc = [&]() -> mlir::LLVM::LLVMFuncOp {
-      llvm::StringRef name("dpcompPurgeContext");
-      auto retType = getVoidType();
-      auto argType = mlir::LLVM::LLVMPointerType::get(getVoidPtrType());
-      auto funcType = mlir::LLVM::LLVMFunctionType::get(retType, argType);
-      return lookupFunc(name, funcType);
-    }();
-
-    auto ctxHandle = [&]() {
-      mlir::OpBuilder::InsertionGuard g(rewriter);
-      rewriter.setInsertionPointToStart(mod.getBody());
-      llvm::StringRef name("context_handle"); // TODO: unique name
-      auto handle = rewriter.create<mlir::LLVM::GlobalOp>(
-          unknownLoc, ctxType, /*isConstant*/ false,
-          mlir::LLVM::Linkage::Internal, name, mlir::Attribute());
-
-      llvm::StringRef cleanupFuncName(".dpcomp_context_cleanup");
-      auto cleanupFunc =
-          mod.lookupSymbol<mlir::LLVM::LLVMFuncOp>(cleanupFuncName);
-      if (!cleanupFunc) {
-        auto cleanupFuncType =
-            mlir::LLVM::LLVMFunctionType::get(getVoidType(), llvm::None);
-        cleanupFunc = rewriter.create<mlir::LLVM::LLVMFuncOp>(
-            unknownLoc, cleanupFuncName, cleanupFuncType);
-        auto block = rewriter.createBlock(&cleanupFunc.getBody());
-        rewriter.setInsertionPointToStart(block);
-        rewriter.create<mlir::LLVM::ReturnOp>(unknownLoc, llvm::None);
-
-        addToGlobalDtors(rewriter, mod, mlir::SymbolRefAttr::get(cleanupFunc),
-                         0);
-      }
-
-      assert(llvm::hasSingleElement(cleanupFunc.getBody()));
-      rewriter.setInsertionPointToStart(&cleanupFunc.getBody().front());
-      mlir::Value addr =
-          rewriter.create<mlir::LLVM::AddressOfOp>(unknownLoc, handle);
-      rewriter.create<mlir::LLVM::CallOp>(unknownLoc, purgeCtxFunc, addr);
-
-      return handle;
-    }();
-
-    auto ctxHandlePtr =
-        rewriter.create<mlir::LLVM::AddressOfOp>(loc, ctxHandle);
-    auto contextSize = getSizeInBytes(loc, ctxStructType, rewriter);
-
-    const mlir::Value takeCtxArgs[] = {
-        ctxHandlePtr,
-        contextSize,
-        initFuncPtr,
-        deinitFuncPtr,
-    };
-    auto ctxPtr =
-        rewriter.create<mlir::LLVM::CallOp>(loc, takeCtxFunc, takeCtxArgs)
-            .getResult(0);
-
-    llvm::SmallVector<mlir::Value> takeCtxResults;
-    takeCtxResults.emplace_back(ctxPtr);
-
-    auto ctxStructPtr =
-        rewriter.create<mlir::LLVM::BitcastOp>(loc, ctxStructPtrType, ctxPtr);
-    auto ctxStruct =
-        rewriter.create<mlir::LLVM::LoadOp>(loc, ctxStructType, ctxStructPtr);
-
-    for (auto i : llvm::seq(0u, resultsCount)) {
-      auto pos = rewriter.getI64ArrayAttr(i);
-      auto res = rewriter.create<mlir::LLVM::ExtractValueOp>(
-          loc, resultTypes[i], ctxStruct, pos);
-      takeCtxResults.emplace_back(res);
-    }
-
-    rewriter.replaceOp(op, takeCtxResults);
-    return mlir::success();
-  }
-};
-
-struct LowerReleaseContextOp
-    : public mlir::ConvertOpToLLVMPattern<plier::ReleaseContextOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  mlir::LogicalResult
-  matchAndRewrite(plier::ReleaseContextOp op,
-                  plier::ReleaseContextOp::Adaptor adaptor,
-                  mlir::ConversionPatternRewriter &rewriter) const override {
-    auto mod = op->getParentOfType<mlir::ModuleOp>();
-    assert(mod);
-
-    auto unknownLoc = rewriter.getUnknownLoc();
-    auto loc = op->getLoc();
-
-    auto lookupFunc = [&](mlir::StringRef name, mlir::Type type) {
-      // TODO: fix and use lookupOrCreateFn
-      if (auto func = mod.lookupSymbol<mlir::LLVM::LLVMFuncOp>(name))
-        return func;
-
-      mlir::OpBuilder::InsertionGuard g(rewriter);
-      rewriter.setInsertionPointToStart(mod.getBody());
-      return rewriter.create<mlir::LLVM::LLVMFuncOp>(
-          unknownLoc, name, type, mlir::LLVM::Linkage::External);
-    };
-
-    auto releaseCtxFunc = [&]() -> mlir::LLVM::LLVMFuncOp {
-      llvm::StringRef name("dpcompReleaseContext");
-      auto voidPtr = getVoidPtrType();
-      auto funcType = mlir::LLVM::LLVMFunctionType::get(voidPtr, voidPtr);
-      return lookupFunc(name, funcType);
-    }();
-
-    rewriter.create<mlir::LLVM::CallOp>(loc, releaseCtxFunc, adaptor.context());
-    rewriter.eraseOp(op);
-    return mlir::success();
-  }
-};
-
-/// Convert operations from the plier_util dialect to the LLVM dialect.
-struct PlierUtilToLLVMPass
-    : public mlir::PassWrapper<PlierUtilToLLVMPass,
-                               mlir::OperationPass<mlir::ModuleOp>> {
-  MLIR_DEFINE_EXPLICIT_INTERNAL_INLINE_TYPE_ID(PlierUtilToLLVMPass)
-
-  PlierUtilToLLVMPass() = default;
-
-  void runOnOperation() override {
-    mlir::Operation *op = getOperation();
-    auto &context = getContext();
-    auto options = getLLVMOptions(context);
-
-    mlir::LLVMTypeConverter typeConverter(&context, options);
-    populateToLLVMAdditionalTypeConversion(typeConverter);
-    mlir::RewritePatternSet patterns(&context);
-
-    patterns.insert<
-        // clang-format off
-        LowerUndef,
-        LowerBuildTuple,
-        LowerRetainOp,
-        ExpandShapeLowering,
-        LowerExtractMemrefMetadataOp,
-        LowerTakeContextOp,
-        LowerReleaseContextOp
-        // clang-format on
-        >(typeConverter);
-
-    mlir::LLVMConversionTarget target(context);
-    target.addLegalOp<mlir::func::FuncOp>();
-    target.addLegalOp<mlir::func::CallOp>();
-    target.addIllegalDialect<plier::PlierUtilDialect>();
-    if (failed(applyPartialConversion(op, target, std::move(patterns))))
-      signalPassFailure();
-  }
-};
-
 // Copypasted from mlir
 struct LLVMLoweringPass
     : public mlir::PassWrapper<LLVMLoweringPass,
@@ -1790,8 +1245,6 @@ struct LLVMLoweringPass
                StringAttr::get(m.getContext(),
                                options.dataLayout.getStringRepresentation()));
   }
-
-private:
 };
 
 static void populateLowerToLlvmPipeline(mlir::OpPassManager &pm) {
@@ -1803,7 +1256,7 @@ static void populateLowerToLlvmPipeline(mlir::OpPassManager &pm) {
   pm.addNestedPass<mlir::func::FuncOp>(std::make_unique<PreLLVMLowering>());
   pm.addNestedPass<mlir::func::FuncOp>(mlir::createConvertMathToLLVMPass());
   pm.addPass(mlir::createConvertMathToLibmPass());
-  pm.addPass(std::make_unique<PlierUtilToLLVMPass>());
+  pm.addPass(plier::createUtilToLLVMPass(&getLLVMOptions));
   pm.addPass(std::make_unique<LLVMLoweringPass>());
   pm.addNestedPass<mlir::LLVM::LLVMFuncOp>(
       std::make_unique<PostLLVMLowering>());
